Add Discrete state class
